Move unrelated functions from zeebe_adapter_base
